+++ src/mineComponents/NavigationBar.tsx
@@ -1,65 +1,360 @@
-import {NavigationMenu, NavigationMenuContent, NavigationMenuItem, NavigationMenuLink, NavigationMenuList, NavigationMenuTrigger, navigationMenuTriggerStyle} from "@/components/ui/navigation-menu";
-import { Link } from "react-router-dom";
-import proveLogo from "../imgs/PROVElogoN.png";
-import cart from "../imgs/cart.svg";
+import { Fragment, useEffect, useState, useContext } from 'react'
+import { Dialog, Popover, Tab, Transition } from '@headlessui/react'
+import { Bars3Icon, ShoppingBagIcon, XMarkIcon } from '@heroicons/react/24/outline'
+import proveLogo from "../imgs/PROVElogoN.png"
+import chupete from "../imgs/chupete.jpg"
+import enzo from "../imgs/enzo.jpg"
 
+import { Link } from 'react-router-dom'
+import { CartContext } from '@/mineComponents/context'
+
+const navigation = {
+  categories: [
+    {
+      id: 'ltw',
+      name: 'LTW',
+      featured: [
+        {
+          name: 'Estilo y Minimalismo',
+          href: '#',
+          imageSrc: chupete,
+          imageAlt: 'Models sitting back to back, wearing Basic Tee in black and bone.',
+        },
+        {
+          name: 'Save Your Broken <3',
+          href: '#',
+          imageSrc: enzo,
+          imageAlt: 'Close up of Basic Tee fall bundle with off-white, ochre, olive, and black tees.',
+        },
+      ],
+      sections: [
+        {
+          id: 'vestir',
+          name: 'Vestir',
+          items: [
+            { name: 'Remeras', href: '#' },
+            { name: 'Pantalones', href: '#' },
+            { name: 'Shorts', href: '#' },
+            { name: 'Etc..', href: '#' },
+          ],
+        },
+      ],
+    },
+    {
+      id: 'basic-collection',
+      name: 'Basic Collection',
+      featured: [
+        {
+          name: 'Ultimos Ingresos',
+          href: '#',
+          imageSrc: 'https://provestoree.com/wp-content/uploads/2022/11/Remera-prove-marron2.jpg',
+          imageAlt: 'Drawstring top with elastic loop closure and textured interior padding.',
+        },
+        {
+          name: 'Nuevos Buzos',
+          href: '#',
+          imageSrc: 'https://provestoree.com/wp-content/uploads/2023/06/Buzo-marron1.jpg',
+          imageAlt:
+            'Three shirts in gray, white, and blue arranged on table with same line drawing of hands and shapes overlapping on front of shirt.',
+        },
+      ],
+      sections: [
+        {
+          id: 'Vestir',
+          name: 'Vestir',
+          items: [
+            { name: 'Buzos', href: '#' },
+            { name: 'Remeras', href: '#' },
+            { name: 'Remerones', href: '#' },
+            { name: 'Remeritas', href: '#' },
+            { name: 'Etc..', href: '#' },
+          ],
+        }
+      ],
+    },
+    
+  ],
+  pages: [
+    { name: 'Op1', href: '#' },
+    { name: 'Op2', href: '#' },
+  ],
+}
+
+function classNames(...classes: string[]) {
+  return classes.filter(Boolean).join(' ')
+}
 
 export function NavigationBar() {
+  const [open, setOpen] = useState(false)
+  const [cantCart, setCantidad] = useState(0)
+  const cartContext = useContext(CartContext);
+  if (cartContext === null) {
+    throw new Error("Error al obtener el contexto del carrito");
+  }
+  const { cart, showCantidad } = cartContext;
+
+  useEffect(() => {
+    setCantidad(showCantidad());
+  },[cart])
+
   return (
+    <div className="bg-white sticky top-0 z-10">
+      {/* Mobile menu */}
+      <Transition.Root show={open} as={Fragment}>
+        <Dialog as="div" className="relative z-40 lg:hidden" onClose={setOpen}>
+          <Transition.Child
+            as={Fragment}
+            enter="transition-opacity ease-linear duration-300"
+            enterFrom="opacity-0"
+            enterTo="opacity-100"
+            leave="transition-opacity ease-linear duration-300"
+            leaveFrom="opacity-100"
+            leaveTo="opacity-0"
+          >
+            <div className="fixed inset-0 bg-black bg-opacity-25" />
+          </Transition.Child>
 
-    <nav className="justify-between flex sticky top-0 bg-white z-10 items-center  border-b border-gray mt-1 ">
-      <button className="md:hidden">
-        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" strokeWidth="2" stroke="currentColor" fill="none" strokeLinecap="round" strokeLinejoin="round" className="ml-7 mr-7">
-          <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
-          <path d="M4 6l16 0"></path>
-          <path d="M4 12l16 0"></path>
-          <path d="M4 18l16 0"></path>
-        </svg>
-      </button>
-      <div className="flex">
-        <Link to="/"><img src={proveLogo} alt="" className="h-7 m-3 ml-7" /></Link>
-        <NavigationMenu className="hidden md:flex">
-          <NavigationMenuList>
-            <Link className={navigationMenuTriggerStyle()} to="/">Inicio</Link>
+          <div className="fixed inset-0 z-40 flex">
+            <Transition.Child
+              as={Fragment}
+              enter="transition ease-in-out duration-300 transform"
+              enterFrom="-translate-x-full"
+              enterTo="translate-x-0"
+              leave="transition ease-in-out duration-300 transform"
+              leaveFrom="translate-x-0"
+              leaveTo="-translate-x-full"
+            >
+              <Dialog.Panel className="relative flex w-full max-w-xs flex-col overflow-y-auto bg-white pb-12 shadow-xl">
+                <div className="flex px-4 pb-2 pt-5">
+                  <button
+                    type="button"
+                    className="relative -m-2 inline-flex items-center justify-center rounded-md p-2 text-gray-400"
+                    onClick={() => setOpen(false)}
+                  >
+                    <span className="absolute -inset-0.5" />
+                    <span className="sr-only">Close menu</span>
+                    <XMarkIcon className="h-6 w-6" aria-hidden="true" />
+                  </button>
+                </div>
 
-            <NavigationMenuItem>
-              <NavigationMenuTrigger>Productos</NavigationMenuTrigger>
-              <NavigationMenuContent>
-                <ul className="grid p-3 md:w-[600px] lg:w-[auto] lg:grid-cols-[.75fr_1fr]">
-                  <li className="row-span-3">
-                    <div className="flex gap-2">
-                      <Link to="/articulos/buzos"
-                        className="flex h-full w-full select-none flex-col justify-end rounded-md bg-gradient-to-b from-muted/50 to-muted items-center no-underline outline-none focus:shadow-md p-2"
-                      >
-                        Buzos
-                      </Link>
-                      <Link to="/articulos/remeras"
-                        className="flex h-full w-full select-none flex-col justify-end rounded-md bg-gradient-to-b from-muted/50 to-muted items-center no-underline outline-none focus:shadow-md p-2"
-                      >
-                        Remeras
-                      </Link>
-                      <Link to="/articulos/shorts"
-                        className="flex h-full w-full select-none flex-col justify-end rounded-md bg-gradient-to-b from-muted/50 to-muted items-center no-underline outline-none focus:shadow-md p-2"
-                      >
-                        Shorts
-                      </Link>
+                {/* Links */}
+                <Tab.Group as="div" className="mt-2">
+                  <div className="border-b border-gray-200">
+                    <Tab.List className="-mb-px flex space-x-8 px-4">
+                      {navigation.categories.map((category) => (
+                        <Tab
+                          key={category.name}
+                          className={({ selected }) =>
+                            classNames(
+                              selected ? 'border-purple-700 text-purple-700' : 'border-transparent text-gray-900',
+                              'flex-1 whitespace-nowrap border-b-2 px-1 py-4 text-base font-medium'
+                            )
+                          }
+                        >
+                          {category.name}
+                        </Tab>
+                      ))}
+                    </Tab.List>
+                  </div>
+                  <Tab.Panels as={Fragment}>
+                    {navigation.categories.map((category) => (
+                      <Tab.Panel key={category.name} className="space-y-10 px-4 pb-8 pt-10">
+                        <div className="grid grid-cols-2 gap-x-4">
+                          {category.featured.map((item) => (
+                            <div key={item.name} className="group relative text-sm">
+                              <div className="aspect-h-1 aspect-w-1 overflow-hidden rounded-lg bg-gray-100 group-hover:opacity-75">
+                                <img src={item.imageSrc} alt={item.imageAlt} className="object-cover object-center" />
+                              </div>
+                              <a href={item.href} className="mt-6 block font-medium text-gray-900">
+                                <span className="absolute inset-0 z-10" aria-hidden="true" />
+                                {item.name}
+                              </a>
+                              <p aria-hidden="true" className="mt-1">
+                                Shop now
+                              </p>
+                            </div>
+                          ))}
+                        </div>
+                        {category.sections.map((section) => (
+                          <div key={section.name}>
+                            <p id={`${category.id}-${section.id}-heading-mobile`} className="font-medium text-gray-900">
+                              {section.name}
+                            </p>
+                            <ul
+                              role="list"
+                              aria-labelledby={`${category.id}-${section.id}-heading-mobile`}
+                              className="mt-6 flex flex-col space-y-6"
+                            >
+                              {section.items.map((item) => (
+                                <li key={item.name} className="flow-root">
+                                  <a href={item.href} className="-m-2 block p-2 text-gray-500">
+                                    {item.name}
+                                  </a>
+                                </li>
+                              ))}
+                            </ul>
+                          </div>
+                        ))}
+                      </Tab.Panel>
+                    ))}
+                  </Tab.Panels>
+                </Tab.Group>
+
+                <div className="space-y-6 border-t border-gray-200 px-4 py-6">
+                  {navigation.pages.map((page) => (
+                    <div key={page.name} className="flow-root">
+                      <a href={page.href} className="-m-2 block p-2 font-medium text-gray-900">
+                        {page.name}
+                      </a>
                     </div>
-                  </li>
-                </ul>
-              </NavigationMenuContent>
-            </NavigationMenuItem>
-            <NavigationMenuItem>
-              <Link className={navigationMenuTriggerStyle()} to="/articles/basic-colection">Basic Colection</Link>
-              <NavigationMenuLink className={navigationMenuTriggerStyle()}>
-                Spring Soon
-              </NavigationMenuLink>
-            </NavigationMenuItem>
-          </NavigationMenuList>
-        </NavigationMenu>
-      </div>
-      <Link to="/cart">
-        <img src={cart} className="w-7 ml-7 mr-10" alt="" />
-      </Link>
-    </nav>
-  );
+                  ))}
+                </div>
+
+              </Dialog.Panel>
+            </Transition.Child>
+          </div>
+        </Dialog>
+      </Transition.Root>
+
+      <header className="relative bg-white">
+        {/* <p className="flex h-10 items-center justify-center bg-purple-700 px-4 text-sm font-medium text-white sm:px-6 lg:px-8">
+          Get free delivery on orders over $100
+        </p> */}
+
+        <nav aria-label="Top" className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
+          <div className="border-b border-gray-200">
+            <div className="flex h-16 items-center">
+              <button
+                type="button"
+                className="relative rounded-md bg-white p-2 text-gray-400 lg:hidden"
+                onClick={() => setOpen(true)}
+              >
+                <span className="absolute -inset-0.5" />
+                <span className="sr-only">Open menu</span>
+                <Bars3Icon className="h-6 w-6" aria-hidden="true" />
+              </button>
+
+              {/* Logo */}
+              <div className="ml-4 flex lg:ml-0">
+              <Link to="/"><img src={proveLogo} alt="" className="h-7 m-3 ml-7" /></Link>
+              </div>
+
+              {/* Flyout menus */}
+              <Popover.Group className="hidden lg:ml-8 lg:block lg:self-stretch">
+                <div className="flex h-full space-x-8">
+                  {navigation.categories.map((category) => (
+                    <Popover key={category.name} className="flex">
+                      {({ open }) => (
+                        <>
+                          <div className="relative flex">
+                            <Popover.Button
+                              className={classNames(
+                                open
+                                  ? 'border-purple-700 text-purple-700'
+                                  : 'border-transparent text-gray-700 hover:text-gray-800',
+                                'relative z-10 -mb-px flex items-center border-b-2 pt-px text-sm font-medium transition-colors duration-200 ease-out'
+                              )}
+                            >
+                              {category.name}
+                            </Popover.Button>
+                          </div>
+
+                          <Transition
+                            as={Fragment}
+                            enter="transition ease-out duration-200"
+                            enterFrom="opacity-0"
+                            enterTo="opacity-100"
+                            leave="transition ease-in duration-150"
+                            leaveFrom="opacity-100"
+                            leaveTo="opacity-0"
+                          >
+                            <Popover.Panel className="absolute inset-x-0 top-full text-sm text-gray-500">
+                              {/* Presentational element used to render the bottom shadow, if we put the shadow on the actual panel it pokes out the top, so we use this shorter element to hide the top of the shadow */}
+                              <div className="absolute inset-0 top-1/2 bg-white shadow" aria-hidden="true" />
+
+                              <div className="relative bg-white">
+                                <div className="mx-auto max-w-7xl px-8">
+                                  <div className="grid grid-cols-2 gap-x-8 gap-y-10 py-16">
+                                    <div className="col-start-2 grid grid-cols-2 gap-x-8">
+                                      {category.featured.map((item) => (
+                                        <div key={item.name} className="group relative text-base sm:text-sm">
+                                          <div className="aspect-h-1 aspect-w-1 overflow-hidden rounded-lg bg-gray-100 group-hover:opacity-75">
+                                            <img
+                                              src={item.imageSrc}
+                                              alt={item.imageAlt}
+                                              className="object-cover object-center"
+                                            />
+                                          </div>
+                                          <a href={item.href} className="mt-6 block font-medium text-gray-900">
+                                            <span className="absolute inset-0 z-10" aria-hidden="true" />
+                                            {item.name}
+                                          </a>
+                                          <p aria-hidden="true" className="mt-1">
+                                            Shop now
+                                          </p>
+                                        </div>
+                                      ))}
+                                    </div>
+                                    <div className="row-start-1 grid grid-cols-3 gap-x-8 gap-y-10 text-sm">
+                                      {category.sections.map((section) => (
+                                        <div key={section.name}>
+                                          <p id={`${section.name}-heading`} className="font-medium text-gray-900">
+                                            {section.name}
+                                          </p>
+                                          <ul
+                                            role="list"
+                                            aria-labelledby={`${section.name}-heading`}
+                                            className="mt-6 space-y-6 sm:mt-4 sm:space-y-4"
+                                          >
+                                            {section.items.map((item) => (
+                                              <li key={item.name} className="flex">
+                                                <a href={item.href} className="hover:text-gray-800">
+                                                  {item.name}
+                                                </a>
+                                              </li>
+                                            ))}
+                                          </ul>
+                                        </div>
+                                      ))}
+                                    </div>
+                                  </div>
+                                </div>
+                              </div>
+                            </Popover.Panel>
+                          </Transition>
+                        </>
+                      )}
+                    </Popover>
+                  ))}
+
+                  {navigation.pages.map((page) => (
+                    <a
+                      key={page.name}
+                      href={page.href}
+                      className="flex items-center text-sm font-medium text-gray-700 hover:text-gray-800"
+                    >
+                      {page.name}
+                    </a>
+                  ))}
+                </div>
+              </Popover.Group>
+
+              <div className="ml-auto flex items-center">
+                </div>
+                <div className="ml-4 flow-root lg:ml-6">
+                  <Link to={"/cart"} className='group -m-2 flex items-center p-2'>
+                       <ShoppingBagIcon
+                      className="h-6 w-6 flex-shrink-0 text-gray-400 group-hover:text-gray-500"
+                      aria-hidden="true"
+                    />
+                    <span className="ml-2 text-sm font-medium text-gray-700 group-hover:text-gray-800">{cantCart}</span>
+                    <span className="sr-only">items in cart, view bag</span>
+                  </Link>
+                </div>
+            </div>
+          </div>
+        </nav>
+      </header>
+    </div>
+  )
 }
